# docs/swagger.yaml
basePath: /
definitions:
  internal_auth.AuthTokens:
    properties:
      access_token:
        type: string
      expires_in:
        description: seconds until access token expires
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  internal_auth.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  internal_auth.ForgotPasswordRequest:
    properties:
      email:
        type: string
    type: object
  internal_auth.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.RefreshRequest:
    properties:
      refresh_token:
        type: string
    type: object
  internal_auth.RegisterRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  internal_auth.RegisterResponse:
    properties:
      message:
        type: string
      user:
        $ref: '#/definitions/internal_auth.UserResponse'
    type: object
  internal_auth.ResendVerificationRequest:
    properties:
      email:
        type: string
    type: object
  internal_auth.ResetPasswordRequest:
    properties:
      new_password:
        type: string
      token:
        type: string
    type: object
  internal_auth.UserResponse:
    properties:
      email:
        type: string
      id:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
  description: A production-ready Go REST API template with authentication, email
    verification, and observability.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Go API Template
  version: "1.0"
paths:
  /auth/forgot-password:
    post:
      consumes:
      - application/json
      description: Send a password reset link to the user's email. Always returns
        success to prevent email enumeration.
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ForgotPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Request password reset
      tags:
      - auth
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user and receive access and refresh tokens
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.AuthTokens'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: User login
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Logout user by revoking refresh token and clearing cookies
      parameters:
      - description: Optional refresh token
        in: body
        name: request
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User logout
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Use a refresh token to get a new access token
      parameters:
      - description: Refresh token
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_auth.AuthTokens'
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Create a new user account with email and password. A verification
        email will be sent.
      parameters:
      - description: Registration credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_auth.RegisterResponse'
        "400":
          description: Invalid request or validation error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
  /auth/resend-verification:
    post:
      consumes:
      - application/json
      description: Send a new verification email to the user. Always returns success
        to prevent email enumeration.
      parameters:
      - description: Email address
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResendVerificationRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "429":
          description: Too many requests
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Resend verification email
      tags:
      - auth
  /auth/reset-password:
    post:
      consumes:
      - application/json
      description: Reset a user's password using a valid reset token
      parameters:
      - description: Reset token and new password
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_auth.ResetPasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid request or token
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Reset password
      tags:
      - auth
  /auth/verify-email:
    get:
      consumes:
      - application/json
      description: Verify a user's email address using the verification token sent
        via email
      parameters:
      - description: Verification token
        in: query
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid, expired, or already used token
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/internal_auth.ErrorResponse'
      summary: Verify email address
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"